Fix migration for feature flag maintenance_mode_status
Before this commit:

Problem 1:
`run make start-cluster NODES=2 LEAVE_PLUGINS_DISABLED=1`

`mnesia:table_info(rabbit_node_maintenance_states, all).`
on both rabbit-1 and rabbit-2 will show
`{disc_copies,[]}` and `{ram_copies,['rabbit-1@myhost']}`

This means the table is only present in RAM on `rabbit-1`.

The problem here is that both `rabbit-1` and `rabbit-2` created
the table independently, but after stopping, joining, and starting
`rabbit-2`, only `rabbit-1` was listed having a copy.

Problem 2:
Same issue as Problem 1, but a different way how we get there.

Create a fresh 3 node cluster on K8s using the rabbitmq/cluster-operator.

The first node being booted will create the `rabbit_node_maintenance_states` table.
The other 2 nodes will print:
```
[warn] <0.228.0> Feature flags: the previous instance of this node must have failed to write the `feature_flags` file at `/var/lib/rabbitmq/mnesia/rabbit@r1-server-0.r1-nodes.default-feature_flags`:
[warn] <0.228.0> Feature flags:   - list of previously enabled feature flags now marked as such: [maintenance_mode_status,
[warn] <0.228.0>                                                                                  quorum_queue,
[warn] <0.228.0>                                                                                  user_limits,
[warn] <0.228.0>                                                                                  virtual_host_metadata]
```
So, the other two nodes do not try to add their copy to the table which leads to
only the booting node to have a copy of the table.

Problem 3:
In Problem 1 + 2, not only did only one node have a copy, but that
copy was also a RAM copy (instead of DISC copy).

What we want:
Disc copies on all nodes after the feature flag has been enabled.

This is achieved in this commit.

Unfortunately, we cannot rely on `maintenance_mode_status_migration(_, _, enable)`
being run on all nodes (see Problem 2).

But we can rely on `maintenance_mode_status_migration(_, _, is_enabled)`
being run on all nodes.

In this commit, we make use of this behaviour by always ensuring that our local node has
a copy of the table if we know that the feature flag is enabled.